added new / delete

+++ src/kernel/include/common.hpp
@@ -17,7 +17,8 @@
 
 #define PACKED __attribute__((packed))
 
-typedef unsigned long long size_t;
+typedef long unsigned int size_t;
+
 typedef unsigned long long uint64_t;
 typedef          long long int64_t;
